add timeout count 1.1 2.1 ...
When the switch is on A (sensor operation), and when the count is anything ending in "1" (e.g. 01.1, 02.1,... x.1) and if after waiting 12 seconds the count does not increase to x.2, then the display should go back to the previous count 'x-1'.0 (i.e reverse the last count. So if the count was 07.1 it should go back to 06.2.)